Factor out SphericalClumpBVH; have it confirm plane crossings exactly
Move the BVH out of ClumpySphericalSpatialGrid into its own standalone
utility class in SKIRT/utils, since it only ever operates on Clump
objects and needs no access to the grid's other private state. Clump
becomes a public nested type of SphericalClumpBVH; ClumpySphericalSpatialGrid
picks it up via a private `using Clump = SphericalClumpBVH::Clump;`
alias to avoid touching every call site.

Drop Clump::numOverlappingCells (and its accessors), since it was only
ever written during setup and never read back; the same information now
lives in a plain local array scoped to setupSelfAfter().

Also have SphericalClumpBVH::clumpsCrossingPlane() perform the exact
sphere/plane test itself and return the intersection circle's radius
alongside each confirmed clump index, rather than a box-based candidate
list the caller had to re-test. The three write_xy/xz/yz call sites in
ClumpySphericalSpatialGrid no longer need any exact-test logic of their
own.
