+++ SKIRT/core/ClumpySphericalSpatialGrid.cpp
@@ -16,7 +16,6 @@
 
 ////////////////////////////////////////////////////////////////////
 
-// Geometric helper functions
 namespace
 {
     // small value, used to progress a path; leave at 1e-10 -- with 1e-11, inaccuracies start to occur
@@ -38,467 +37,6 @@ namespace
         }
     }
 }
-
-////////////////////////////////////////////////////////////////////
-
-namespace
-{
-    // Number of candidate split planes ("bins") evaluated per axis during the SAH search
-    // (binned SAH, Wald & Havran); O(N) work per node rather than O(N log N) for an exact
-    // search, at negligible cost to tree quality.
-    constexpr int NumBins = 16;
-
-    // A leaf never holds more clumps than this, regardless of what SAH suggests. Bounds the
-    // cost of the linear scan at each leaf, and guarantees the build terminates even when
-    // many clumps have coincident or near-coincident centers.
-    constexpr int MaxLeafSize = 4;
-}
-
-////////////////////////////////////////////////////////////////////
-
-// Private utility class for organizing spheres in a data structure that allows efficient queries.
-// The implementation employs a linearized Bounding Volume Hierarchy (BVH) that is bulk-loaded
-// using a surface area heuristic (SAH) for optimal balance. Construction of the data structure
-// runs in a single serial thread. Because the data structure does not change after initial
-// construction, all queries are thread-safe.
-class ClumpySphericalSpatialGrid::BVH
-{
-private:
-    class Node
-    {
-    public:
-        Box box;         // union of the bounding boxes of all clumps in this node's subtree
-        int left;        // index into _nodes of the left child, or -1 if this is a leaf
-        int right;       // index into _nodes of the right child, or -1 if this is a leaf
-        int firstIndex;  // for a leaf: start of this leaf's range in _index
-        int numIndices;  // for a leaf: number of clumps in this leaf; 0 for an interior node
-
-        Node(const Box& box_, int left_, int right_, int firstIndex_, int numIndices_)
-            : box(box_), left(left_), right(right_), firstIndex(firstIndex_), numIndices(numIndices_)
-        {}
-
-        bool isLeaf() const { return numIndices > 0; }
-    };
-
-    // flat array of BVH nodes; index 0 is the root, meaningful only when _numEntities > 0
-    vector<Node> _nodes;
-
-    // entity indices, reordered during the build so that the clumps of any single leaf
-    // occupy the contiguous range [firstIndex, firstIndex+numIndices) of this array
-    vector<int> _index;
-
-    // pointer to array with clumps, passed on when bulk-loading the BVH
-    const vector<Clump>* _clumps;
-
-    // Recursively builds the BVH using binned SAH over the clumps referenced by
-    // _index[begin,end), reordering that sub-range in place. Appends the newly created
-    // node(s) to _nodes and returns the index of the node representing this sub-range.
-    int buildRecursive(int begin, int end)
-    {
-        // aggregate bounding box for this sub-range: the union of the entity boxes referenced
-        // by _index[begin,end).
-        Box box = (*_clumps)[_index[begin]].bounds();
-        for (int i = begin + 1; i != end; ++i) box.extend((*_clumps)[_index[i]].bounds());
-
-        int numPrims = end - begin;
-        int nodeIndex = static_cast<int>(_nodes.size());
-
-        auto makeLeaf = [&]() {
-            _nodes.emplace_back(box, -1, -1, begin, numPrims);
-            return nodeIndex;
-        };
-
-        if (numPrims <= MaxLeafSize) return makeLeaf();
-
-        // centroid bounding box, used to pick a split axis and the bin boundaries along it
-        Position c0 = (*_clumps)[_index[begin]].center();
-        double cmin[3] = {c0.x(), c0.y(), c0.z()};
-        double cmax[3] = {c0.x(), c0.y(), c0.z()};
-        for (int i = begin + 1; i != end; ++i)
-        {
-            Position c = (*_clumps)[_index[i]].center();
-            double v[3] = {c.x(), c.y(), c.z()};
-            for (int a = 0; a != 3; ++a)
-            {
-                cmin[a] = min(cmin[a], v[a]);
-                cmax[a] = max(cmax[a], v[a]);
-            }
-        }
-        int axis = 0;
-        double extentOnAxis = cmax[0] - cmin[0];
-        for (int a = 1; a != 3; ++a)
-        {
-            double e = cmax[a] - cmin[a];
-            if (e > extentOnAxis)
-            {
-                extentOnAxis = e;
-                axis = a;
-            }
-        }
-
-        // all centroids coincide -> no split can separate the clumps; make a leaf
-        // (also protects the binning step below from a division by zero)
-        if (extentOnAxis <= 0.0) return makeLeaf();
-
-        auto centroidAxis = [axis](Position c) { return axis == 0 ? c.x() : axis == 1 ? c.y() : c.z(); };
-
-        // bin the clumps of [begin,end) along the chosen axis
-        struct Bin
-        {
-            int count = 0;
-            Box box;
-            bool hasBox = false;
-        };
-        vector<Bin> bins(NumBins);
-        double lo = cmin[axis], hi = cmax[axis];
-        auto binIndexOf = [&](double v) {
-            int b = static_cast<int>(NumBins * (v - lo) / (hi - lo));
-            return min(max(b, 0), NumBins - 1);
-        };
-        for (int i = begin; i != end; ++i)
-        {
-            const Clump& c = (*_clumps)[_index[i]];
-            int b = binIndexOf(centroidAxis(c.center()));
-            Box eb = c.bounds();
-            if (bins[b].hasBox)
-                bins[b].box.extend(eb);
-            else
-            {
-                bins[b].box = eb;
-                bins[b].hasBox = true;
-            }
-            bins[b].count++;
-        }
-
-        // running unions/counts from the left and from the right, used to evaluate the SAH
-        // cost of splitting after each bin boundary without re-scanning for every candidate
-        vector<Box> prefixBox(NumBins);
-        vector<int> prefixCount(NumBins);
-        vector<Box> suffixBox(NumBins);
-        vector<int> suffixCount(NumBins);
-        {
-            bool has = false;
-            Box running;
-            int runningCount = 0;
-            for (int b = 0; b != NumBins; ++b)
-            {
-                if (bins[b].hasBox)
-                {
-                    if (has)
-                        running.extend(bins[b].box);
-                    else
-                    {
-                        running = bins[b].box;
-                        has = true;
-                    }
-                }
-                runningCount += bins[b].count;
-                prefixBox[b] = running;
-                prefixCount[b] = runningCount;
-            }
-        }
-        {
-            bool has = false;
-            Box running;
-            int runningCount = 0;
-            for (int b = NumBins - 1; b >= 0; --b)
-            {
-                if (bins[b].hasBox)
-                {
-                    if (has)
-                        running.extend(bins[b].box);
-                    else
-                    {
-                        running = bins[b].box;
-                        has = true;
-                    }
-                }
-                runningCount += bins[b].count;
-                suffixBox[b] = running;
-                suffixCount[b] = runningCount;
-            }
-        }
-
-        // evaluate the SAH cost of splitting right after bin b, for every interior bin boundary,
-        // and keep the cheapest; cost is expressed relative to the parent's surface area so it
-        // can be compared directly against the cost of not splitting at all
-        double parentArea = box.surfaceArea();
-        double bestCost = std::numeric_limits<double>::infinity();
-        int bestSplit = -1;
-        for (int b = 0; b != NumBins - 1; ++b)
-        {
-            if (prefixCount[b] == 0 || suffixCount[b + 1] == 0) continue;
-            double cost =
-                (prefixCount[b] * prefixBox[b].surfaceArea() + suffixCount[b + 1] * suffixBox[b + 1].surfaceArea())
-                / parentArea;
-            if (cost < bestCost)
-            {
-                bestCost = cost;
-                bestSplit = b;
-            }
-        }
-
-        // no viable split, or splitting is not cheaper than a leaf -> make a leaf
-        double leafCost = static_cast<double>(numPrims);
-        if (bestSplit < 0 || bestCost >= leafCost) return makeLeaf();
-
-        double splitPos = lo + (hi - lo) * (bestSplit + 1) / NumBins;
-        auto midIt = std::partition(_index.begin() + begin, _index.begin() + end,
-                                    [&](int m) { return centroidAxis((*_clumps)[m].center()) < splitPos; });
-        int mid = static_cast<int>(midIt - _index.begin());
-        if (mid == begin || mid == end) mid = (begin + end) / 2;  // guard against a degenerate partition
-
-        // reserve this node's slot now; children are appended afterwards and get higher
-        // indices, so we come back and fill in left/right once the recursion returns
-        _nodes.emplace_back(box, -1, -1, 0, 0);
-        int leftIndex = buildRecursive(begin, mid);
-        int rightIndex = buildRecursive(mid, end);
-        _nodes[nodeIndex].left = leftIndex;
-        _nodes[nodeIndex].right = rightIndex;
-        return nodeIndex;
-    }
-
-public:
-    // Bulk-loads the specified clumps into the BVH
-    void loadClumps(const vector<Clump>& clumps)
-    {
-        _clumps = &clumps;
-        int numClumps = static_cast<int>(clumps.size());
-
-        _nodes.clear();
-        _index.resize(numClumps);
-        for (int m = 0; m != numClumps; ++m) _index[m] = m;
-
-        if (numClumps > 0)
-        {
-            // rough heuristic upper bound on the final node count (a binary tree with roughly
-            // numEntities/MaxLeafSize leaves has on the order of twice as many nodes total)
-            _nodes.reserve(2 * numClumps / MaxLeafSize + 1);
-            buildRecursive(0, numClumps);
-        }
-    }
-
-    // Returns the indices of a maximal subset of mutually non-overlapping clumps, built
-    // greedily in order of increasing index: clump 0 is always kept, and each subsequent
-    // clump is kept unless it overlaps a clump already kept. Uses the BVH already built
-    // over ALL clumps (including ones that will be rejected here) purely to prune the
-    // overlap search down to nearby candidates via box intersection, rather than checking
-    // against every previously-accepted clump directly (which would be O(M^2) overall).
-    vector<int> allDisjointClumps() const
-    {
-        vector<int> result;
-        int numClumps = static_cast<int>(_index.size());
-        if (numClumps == 0) return result;
-
-        // accepted[m] becomes true once entity m has been added to the result; a plain
-        // vector<char> is used instead of vector<bool> for fast, unambiguous random-access.
-        vector<char> accepted(numClumps, 0);
-        accepted[0] = 1;
-        result.push_back(0);
-
-        // For each remaining clump, we need to know whether it overlaps ANY already-accepted clump
-        vector<int> stack;
-        for (int i = 1; i != numClumps; ++i)
-        {
-            const Clump& ci = (*_clumps)[i];
-            Box queryBox = ci.bounds();
-            bool conflict = false;
-
-            stack.clear();
-            stack.push_back(0);
-            while (!stack.empty() && !conflict)
-            {
-                int nodeIndex = stack.back();
-                stack.pop_back();
-                const Node& node = _nodes[nodeIndex];
-                if (!node.box.intersects(queryBox)) continue;
-
-                if (node.isLeaf())
-                {
-                    for (int k = node.firstIndex; k != node.firstIndex + node.numIndices; ++k)
-                    {
-                        int m = _index[k];
-                        if (!accepted[m]) continue;
-                        const Clump& cm = (*_clumps)[m];
-                        if (Quadrics::doSpheresOverlap(ci.center(), ci.radius(), cm.center(), cm.radius()))
-                        {
-                            conflict = true;
-                            break;
-                        }
-                    }
-                }
-                else
-                {
-                    stack.push_back(node.left);
-                    stack.push_back(node.right);
-                }
-            }
-
-            if (!conflict)
-            {
-                accepted[i] = 1;
-                result.push_back(i);
-            }
-        }
-        return result;
-    }
-
-    // Returns the index of any clump containing the given position, or -1 if none does.
-    // No ordering is needed (clumps are non-overlapping by the time this is used in
-    // practice), so a plain box-pruned depth-first search returning on the first hit is
-    // the fastest option.
-    int anyClumpContaining(Vec bfr) const
-    {
-        if (_nodes.empty()) return -1;
-
-        // use a thread_local stack to avoid allocations between consecutibe queries
-        thread_local vector<int> stack;
-        stack.clear();
-        stack.push_back(0);
-
-        while (!stack.empty())
-        {
-            int nodeIndex = stack.back();
-            stack.pop_back();
-            const Node& node = _nodes[nodeIndex];
-            if (!node.box.contains(bfr)) continue;
-
-            if (node.isLeaf())
-            {
-                for (int k = node.firstIndex; k != node.firstIndex + node.numIndices; ++k)
-                {
-                    int m = _index[k];
-                    const Clump& c = (*_clumps)[m];
-                    if (Quadrics::isPositionInSphere(bfr, c.center(), c.radius())) return m;
-                }
-            }
-            else
-            {
-                stack.push_back(node.left);
-                stack.push_back(node.right);
-            }
-        }
-        return -1;
-    }
-
-    // Returns the index of the nearest clump intersected by the ray (r0,k) at a forward
-    // distance strictly between 0 and sBest, or -1 if there is none. On success, sBest is
-    // set to the entry distance of that specific clump.
-    //
-    // This is a nearest-first, depth-first traversal with branch-and-bound pruning: a node's
-    // ray-box entry distance (from Box::intersects) is a valid lower bound on the entry
-    // distance of any clump inside it, so a node whose entry distance is not smaller than the
-    // best confirmed hit so far can be skipped outright. Visiting the geometrically nearer
-    // child first tends to tighten that bound quickly. Unlike an ordered multi-hit iterator,
-    // this only ever needs the single nearest hit, since the segment generator re-queries
-    // fresh from its new position after every segment.
-    int nearestClumpAlongRay(Position r0, Direction k, double& sBest) const
-    {
-        if (_nodes.empty()) return -1;
-
-        int bestIndex = -1;
-
-        struct StackEntry
-        {
-            int nodeIndex;
-            double boxEntry;
-        };
-        thread_local vector<StackEntry> stack;  // thread_local to avoid reallocations on repeated use
-        stack.clear();
-
-        double smin, smax;
-        if (_nodes[0].box.intersects(r0, k, smin, smax) && smin < sBest) stack.push_back({0, smin});
-
-        while (!stack.empty())
-        {
-            StackEntry top = stack.back();
-            stack.pop_back();
-            if (top.boxEntry >= sBest) continue;  // superseded by a tighter bound found meanwhile
-
-            const Node& node = _nodes[top.nodeIndex];
-            if (node.isLeaf())
-            {
-                for (int i = node.firstIndex; i != node.firstIndex + node.numIndices; ++i)
-                {
-                    int m = _index[i];
-                    const Clump& c = (*_clumps)[m];
-                    double s0 = Quadrics::firstIntersectionSphere(r0, k, c.center(), c.radius());
-                    if (s0 > 0. && s0 < sBest)
-                    {
-                        sBest = s0;
-                        bestIndex = m;
-                    }
-                }
-            }
-            else
-            {
-                double sminL, smaxL, sminR, smaxR;
-                bool hitL = _nodes[node.left].box.intersects(r0, k, sminL, smaxL) && sminL < sBest;
-                bool hitR = _nodes[node.right].box.intersects(r0, k, sminR, smaxR) && sminR < sBest;
-
-                // push the farther child first so the nearer one is popped (and processed) first
-                if (hitL && hitR)
-                {
-                    if (sminL <= sminR)
-                    {
-                        stack.push_back({node.right, sminR});
-                        stack.push_back({node.left, sminL});
-                    }
-                    else
-                    {
-                        stack.push_back({node.left, sminL});
-                        stack.push_back({node.right, sminR});
-                    }
-                }
-                else if (hitL)
-                    stack.push_back({node.left, sminL});
-                else if (hitR)
-                    stack.push_back({node.right, sminR});
-            }
-        }
-
-        return bestIndex;
-    }
-
-    // Returns the indices of the clumps whose bounding box may cross the coordinate plane
-    // where coordinate 'axis' (0=x, 1=y, 2=z) equals 'value'. Runs once at plotting time, so
-    // a plain box-pruned traversal (no thread_local reuse) is fine. The result can include a
-    // few clumps whose bounding box straddles the plane but whose actual sphere doesn't quite
-    // reach it; the caller is expected to apply the exact test itself, which it needs anyway
-    // to get the intersection circle's radius.
-    vector<int> clumpsCrossingPlane(int axis, double value) const
-    {
-        vector<int> result;
-        if (_nodes.empty()) return result;
-
-        auto straddles = [axis, value](const Box& b) {
-            double lo = axis == 0 ? b.xmin() : axis == 1 ? b.ymin() : b.zmin();
-            double hi = axis == 0 ? b.xmax() : axis == 1 ? b.ymax() : b.zmax();
-            return lo <= value && value <= hi;
-        };
-
-        vector<int> stack;
-        stack.push_back(0);
-        while (!stack.empty())
-        {
-            int nodeIndex = stack.back();
-            stack.pop_back();
-            const Node& node = _nodes[nodeIndex];
-            if (!straddles(node.box)) continue;
-
-            if (node.isLeaf())
-            {
-                for (int k = node.firstIndex; k != node.firstIndex + node.numIndices; ++k) result.push_back(_index[k]);
-            }
-            else
-            {
-                stack.push_back(node.left);
-                stack.push_back(node.right);
-            }
-        }
-        return result;
-    }
-};
 
 ////////////////////////////////////////////////////////////////////
 
@@ -575,7 +113,7 @@ void ClumpySphericalSpatialGrid::setupSelfAfter()
     }
 
     // build a BVH over all clumps inside the domain, including mutually overlapping ones
-    _bvh = new BVH;
+    _bvh = new SphericalClumpBVH;
     log->info("Constructing bounding volume hierarchy for " + std::to_string(_clumps.size()) + " clumps...");
     _bvh->loadClumps(_clumps);
 
@@ -643,12 +181,14 @@ void ClumpySphericalSpatialGrid::setupSelfAfter()
     };
 
     int numStraddling = 0;
+    // number of structured cells overlapped by each clump; only needed during setup, for logging
+    vector<int> numOverlappingCellsPerClump(_numClumps, 0);
     log->info("Estimating structured-cell volumes overlapped by " + std::to_string(_numClumps) + " clumps (at least "
               + std::to_string(K) + " samples per clump)...");
     vector<std::pair<int, int>> hits;  // reused across clumps to avoid reallocating on every iteration
     for (int ci = 0; ci != _numClumps; ++ci)
     {
-        Clump& clump = _clumps[ci];
+        const Clump& clump = _clumps[ci];
         double clumpVolume = Quadrics::volumeSphere(clump.radius());
 
         // phase 1: draw K samples and discover how many distinct cells are touched
@@ -666,7 +206,7 @@ void ClumpySphericalSpatialGrid::setupSelfAfter()
         }
 
         int numOverlappingCells = static_cast<int>(hits.size());
-        clump.setNumOverlappingCells(numOverlappingCells);
+        numOverlappingCellsPerClump[ci] = numOverlappingCells;
         if (numOverlappingCells > 1) numStraddling++;
 
         for (const auto& hit : hits)
@@ -685,7 +225,7 @@ void ClumpySphericalSpatialGrid::setupSelfAfter()
     log->info("  Rejected (outside domain):  " + std::to_string(numOutside));
     log->info("  Rejected (overlapping):     " + std::to_string(numOverlapping));
     log->info("  Remaining in spatial grid:  " + std::to_string(_numClumps));
-    log->info("  Straddling a structured cell wall: " + std::to_string(numStraddling));
+    log->info("  Straddling a cell wall:     " + std::to_string(numStraddling));
 
     // ---- epsilon used to progress a path, small relative to both the mesh and the smallest clump ----
 
@@ -1057,12 +597,10 @@ void ClumpySphericalSpatialGrid::write_xy(SpatialGridPlotFile* outfile) const
         outfile->writeLine(_rv[0] * cos(phi), _rv[0] * sin(phi), _rv[_Nr] * cos(phi), _rv[_Nr] * sin(phi));
 
     // spherical clumps crossing the xy plane
-    for (int m : _bvh->clumpsCrossingPlane(2, 0.))
+    for (const auto& hit : _bvh->clumpsCrossingPlane(2, 0.))
     {
-        const Clump& c = _clumps[m];
-        double circleRadius;
-        if (Quadrics::sphereIntersectsPlane(c.radius(), c.center().z(), circleRadius))
-            outfile->writeCircle(c.center().x(), c.center().y(), circleRadius);
+        const Clump& c = _clumps[hit.first];
+        outfile->writeCircle(c.center().x(), c.center().y(), hit.second);
     }
 }
 
@@ -1074,12 +612,10 @@ void ClumpySphericalSpatialGrid::write_xz(SpatialGridPlotFile* outfile) const
     writeMeridionalStructure(outfile, _rv, _thetav);
 
     // spherical clumps crossing the xz plane
-    for (int m : _bvh->clumpsCrossingPlane(1, 0.))
+    for (const auto& hit : _bvh->clumpsCrossingPlane(1, 0.))
     {
-        const Clump& c = _clumps[m];
-        double circleRadius;
-        if (Quadrics::sphereIntersectsPlane(c.radius(), c.center().y(), circleRadius))
-            outfile->writeCircle(c.center().x(), c.center().z(), circleRadius);
+        const Clump& c = _clumps[hit.first];
+        outfile->writeCircle(c.center().x(), c.center().z(), hit.second);
     }
 }
 
@@ -1093,12 +629,10 @@ void ClumpySphericalSpatialGrid::write_yz(SpatialGridPlotFile* outfile) const
     writeMeridionalStructure(outfile, _rv, _thetav);
 
     // spherical clumps crossing the yz plane
-    for (int m : _bvh->clumpsCrossingPlane(0, 0.))
+    for (const auto& hit : _bvh->clumpsCrossingPlane(0, 0.))
     {
-        const Clump& c = _clumps[m];
-        double circleRadius;
-        if (Quadrics::sphereIntersectsPlane(c.radius(), c.center().x(), circleRadius))
-            outfile->writeCircle(c.center().y(), c.center().z(), circleRadius);
+        const Clump& c = _clumps[hit.first];
+        outfile->writeCircle(c.center().y(), c.center().z(), hit.second);
     }
 }
